In neomedia, converts RTP payload type from int to byte. Also fixes and adds javadocs.

+++ src/net/java/sip/communicator/impl/neomedia/format/MediaFormatImpl.java
@@ -310,7 +310,7 @@ public abstract class MediaFormatImpl<T extends Format>
      * Control"; otherwise, {@link #RTP_PAYLOAD_TYPE_UNKNOWN}
      * @see MediaFormat#getRTPPayloadType()
      */
-    public int getRTPPayloadType()
+    public byte getRTPPayloadType()
     {
         return MediaUtils.getRTPPayloadType(getJMFEncoding(), getClockRate());
     }
